[ANDROAPP-5469] Extract FieldProvider from Form.kt
